change to long

+++ CombinationGeneratorAPI/CombinationGeneratorAPI.Application/DTOs/GetAllResponse.cs
@@ -1,3 +1,3 @@
 namespace CombinationGeneratorAPI.Application.DTOs;
 public record PermutationItem(int[] Permutation, string Index);
-public record GetAllResponse(IEnumerable<PermutationItem> Permutations, bool HasMore, int TotalPages);
+public record GetAllResponse(IEnumerable<PermutationItem> Permutations, bool HasMore, long TotalPages);
